Standalone functions (#99)
* memoizeAsyncify

* fixing lint

* after

* beforify

* cancel-prev

* debounce

* fixing build

* delay

* delegate

* exec-time

* memoize

* multi-dispatch

* on-error

* rate-limit

* retry

* fix build

* throttle

* throttle-async

* index fix

* fixing build

* timeout

* fixing build

* fixing build

* fixing covarage
