+++ src/after/after.index.ts
@@ -1,2 +1,3 @@
 export {after} from './after';
+export {afterify} from './afterify';
 export {AfterFunc, AfterConfig, AfterParams} from './after.model';
