+++ src/before/before.index.ts
@@ -1,2 +1,3 @@
 export {before} from './before';
+export {beforify} from './beforify';
 export {BeforeConfig} from './before.model';
