+++ src/cancel-previous/cancel-previous.index.ts
@@ -1,2 +1,4 @@
-export {cancelPrevious, CanceledPromise} from './cancel-previous';
+export {cancelPreviousify} from './cancel-previousify';
+export {CanceledPromise} from './canceled-promise';
+export {cancelPrevious} from './cancel-previous';
 export * from './cancel-previous.model';
