+++ src/cancel-previous/canceled-promise.ts
@@ -1,0 +1,7 @@
+export class CanceledPromise extends Error {
+  constructor() {
+    super('canceled');
+
+    Object.setPrototypeOf(this, CanceledPromise.prototype);
+  }
+}
