+++ src/debounce/debounce.index.ts
@@ -1,1 +1,2 @@
+export {debouncify} from './debouncify';
 export {debounce} from './debounce';
