+++ src/delay/delay.index.ts
@@ -1,1 +1,2 @@
 export {delay} from './delay';
+export {delayfy} from './delayfy';
