+++ src/memoize-async/memoize-async.index.ts
@@ -1,2 +1,3 @@
 export {memoizeAsync} from './memoize-async';
+export {memoizeAsyncify} from './memoize-asyncify';
 export {AsyncCache, AsyncMemoizable, AsyncMemoizeConfig} from './memoize-async.model';
