+++ src/memoize/memoize.index.ts
@@ -1,2 +1,3 @@
 export {memoize} from './memoize';
+export {memoizify} from './memoizify';
 export {Cache, Memoizable, MemoizeConfig} from './memoize.model';
