+++ src/memoize/memoize.ts
@@ -1,62 +1,18 @@
 import {Memoizable, MemoizeConfig} from './memoize.model';
 import {Method} from '../common/model/common.model';
-import {TaskExec} from '../common/tesk-exec/task-exec';
+import {memoizify} from './memoizify';
 
 export function memoize<T = any, D = any>(): Memoizable<T, D>;
 export function memoize<T = any, D = any>(config: MemoizeConfig<T, D>): Memoizable<T, D>;
 export function memoize<T = any, D = any>(expirationTimeMs: number): Memoizable<T, D>;
 export function memoize<T = any, D = any>(input?: MemoizeConfig<T, D> | number): Memoizable<T, D> {
-  const defaultConfig: MemoizeConfig<any, D> = {
-    cache: new Map<string, D>(),
-  };
-  const runner = new TaskExec();
-
   return (
     target: T,
     propertyName: keyof T,
     descriptor: TypedPropertyDescriptor<Method<D>>,
   ): TypedPropertyDescriptor<Method<D>> => {
-    let resolvedConfig = {
-      ...defaultConfig,
-    } as MemoizeConfig<T, D>;
-
-    if (typeof input === 'number') {
-      resolvedConfig.expirationTimeMs = input;
-    } else {
-      resolvedConfig = {
-        ...resolvedConfig,
-        ...input,
-      };
-    }
-
     if (descriptor.value) {
-      const originalMethod = descriptor.value;
-      descriptor.value = function (...args: any[]): D {
-        let key;
-        const keyResolver = typeof resolvedConfig.keyResolver === 'string'
-          ? this[resolvedConfig.keyResolver].bind(this)
-          : resolvedConfig.keyResolver;
-
-        if (keyResolver) {
-          key = keyResolver(...args);
-        } else {
-          key = JSON.stringify(args);
-        }
-
-        if (!resolvedConfig.cache.has(key)) {
-          const response = originalMethod.apply(this, args);
-
-          if (resolvedConfig.expirationTimeMs !== undefined) {
-            runner.exec(() => {
-              resolvedConfig.cache.delete(key);
-            }, resolvedConfig.expirationTimeMs);
-          }
-
-          resolvedConfig.cache.set(key, response);
-        }
-
-        return resolvedConfig.cache.get(key);
-      };
+      descriptor.value = memoizify<D>(descriptor.value, input as any);
 
       return descriptor;
     }
