+++ src/multi-dispatch/multi-dispatch.index.ts
@@ -1,1 +1,2 @@
 export {multiDispatch} from './multi-dispatch';
+export {multiDispatchify} from './multi-dispatchify';
