+++ src/on-error/on-error.index.ts
@@ -1,2 +1,3 @@
+export {onErrorify} from './on-errorify';
 export {onError} from './on-error';
 export * from './on-error.model';
