+++ src/rate-limit/rate-limit.index.ts
@@ -1,2 +1,3 @@
+export {rateLimitify} from './rate-limitify';
 export {rateLimit} from './rate-limit';
 export * from './rate-limit.model';
