+++ src/throttle-async/throttle-async.index.ts
@@ -1,1 +1,2 @@
+export {throttleAsyncify} from './throttle-asyncify';
 export {throttleAsync} from './throttle-async';
