+++ src/throttle-async/throttle-async.ts
@@ -1,19 +1,14 @@
 import {AsyncMethod, Decorator} from '../common/model/common.model';
-import {ThrottleAsyncExecutor} from './throttle-async-executor';
+import {throttleAsyncify} from './throttle-asyncify';
 
-export function throttleAsync<T = any, D = any>(parallelCalls = 1): Decorator<T> {
+export function throttleAsync<T = any, D = any>(parallelCalls?: number): Decorator<T> {
   return (
     target: T,
     propertyName: keyof T,
     descriptor: TypedPropertyDescriptor<AsyncMethod<any>>,
   ): TypedPropertyDescriptor<AsyncMethod<D>> => {
     if (descriptor.value) {
-      const originalMethod = descriptor.value;
-      const executor = new ThrottleAsyncExecutor(originalMethod, parallelCalls);
-
-      descriptor.value = function (...args: any[]): Promise<D> {
-        return executor.exec(this, args);
-      };
+      descriptor.value = throttleAsyncify(descriptor.value, parallelCalls);
 
       return descriptor;
     }
