+++ src/throttle/throttle.index.ts
@@ -1,1 +1,2 @@
+export {throttlify} from './throttlify';
 export {throttle} from './throttle';
