+++ src/timeout/timeout-error.ts
@@ -1,0 +1,7 @@
+export class TimeoutError extends Error {
+  constructor(ms: number) {
+    super(`timeout occurred after ${ms}`);
+
+    Object.setPrototypeOf(this, TimeoutError.prototype);
+  }
+}
