+++ src/timeout/timeout.index.ts
@@ -1,1 +1,3 @@
+export {TimeoutError} from './timeout-error';
+export {timeoutify} from './timeoutify';
 export {timeout} from './timeout';
